Reduce the size of Value and Value_View structs

+++ source.h
@@ -109,7 +109,7 @@ static PRELUDE_NO_DISCARD Value *
 token_parse_expression(
   Execution_Context *context,
   Value_View view,
-  u64 *out_match_length,
+  u32 *out_match_length,
   const Token_Pattern *end_pattern
 );
